# openapi/user-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/assessment/attempt/{discipline}": {
      "get": {
        "operationId": "AssessmentController_attempt",
        "parameters": [
          {
            "name": "discipline",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "BJJ_GI",
                "NO_GI_GRAPPLING"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessment"
        ]
      }
    },
...
    "/assessment/roadmap-items/{id}": {
      "get": {
        "operationId": "AssessmentController_details",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessment"
        ]
      },
      "patch": {
        "operationId": "AssessmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoadmapItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessment"
        ]
      }
    },
...
  },
  "info": {
    "title": "MATIQ User API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateRoadmapItemDto": {
        "type": "object",
        "properties": {
          "isHidden": {
            "type": "boolean"
          },
          "direction": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ]
          },
          "completed": {
            "type": "boolean"
          },
          "selectedSkillKeys": {
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}